Migrate logout endpoint, integration tests
